requirements.txt
Flask
feedparser
requests
google-cloud-secret-manager
